+++ src/lorem.js
@@ -1,4 +1,4 @@
-import _ from "lodash";
+import R from "ramda";
 export const LOREM = "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
 const LOREM_WORDS = LOREM.split(" ");
 
@@ -10,7 +10,7 @@ const LOREM_WORDS = LOREM.split(" ");
 
 export default (totalWords) => {
   let result = LOREM_WORDS;
-  if (_.isNumber(totalWords)) {
+  if (R.is(Number, totalWords)) {
     // A specific number of words required.
     if (totalWords < 1) { totalWords = 1; }
     result = [];